I finished a rough outline of index.html
I added the rest of the website including and below the div class="content". I created containers for each square of content like image-subtext-container, each with their own image and subtext container.
Then i had a div class="quote-container" within the content class with two ID p elements. This part of the html file I might have to change in order to get a similar format to the webpage being referenced.
The next container within the content class is the div class="sign-up-container" with a box within to container that seperates the context of the sign up and the button itself, both within their own division to use flex-box.
That is the end of the class=content div. The last part was a div="footer" with a p element for copy right.