One addition to html and header start to style.css
I made one addition to index.html by adding an id value to the header logo. I gave it a header-logo value for the id.

As for the style.css

I have began making changes to class='header'. I maked the backgoround color: #1F2937. I used flexbox to ensure the content are spaced around and will move as the view changes.
I also got rid of the bullet points of the links and the underline. I gave them 0 padding and 0 margin as they come with default values for that. I did however make sure there was a gap of 10px between them.

The body was also made a flexbox as well but in column. Along with margin auto and padding auto.