Sub Header completed in style.css
The changes I made were to body .sub-header, .image-container, left-side-of-sub-header, and I created a new class selector fo just the h1 element.

The changes I made to the boyd is I added flex: 1; I will remove it if I don't see any affects to the content. As my goal is to ensure that that sections are all individually the same height.

The changes I made to sub-header were that I did display: flex;, justify-content: center;, algn-items: center; I gave the content a gap of 18px; and finally I gave it a height of 350px;.

For .left-side-of-header I gave it a width of 500 px and a height of 200px;. I also gave it flexbox properties with flex-direction column since the button already has a width so it won't grow to a box. I also did justify-content.

The new selector for h1 was added because I found out that h1 has their own margin already embedded throwing off the lining of the title and image. So I got rid of the margin finally aligning the sections.