rough draft of the some random information container section
The changes I made to index.html is that I took out the h2 element out the image-subtext container and put it right below the div of content. Then I changed the name of image-subtext-container to image-subtext-containers and changed made it mulitple divisions within a new div container called some-random-info-container. The scoping that was added within the some-random-information container was that the image div no longer contained the p element within and is the same level scope as the div in order to create two containers within the image-subtext-containers. I did this for a total of 4 times. That was all changes I did for index.html

The changes I did for style.css is addition of the new class selectors and element select to work on the some random information sections.

I added .content with flexbox property and gave it a direction of column

I made the h2 element centered with the property of text-align.

The new div that I created named .random-information-container. I gave it the flexbox propety with align and justify content : center;

Within the 4 div element "image-subtext-containers" I gave them a flexbox with the directions of column so that the images will be above the subtext and will be align perfectly. I also gave the justify-align content : center;

The div class="image" was given a width, height of 100px; Along with a display: flex; since there is multiple of them. I will do some testing to see if they need it or if the parent element is enough for it. Other than than that I also gave it flex-direction of column. That too I will change if the parent element covers it with the cascade. The other property values that will stay is the border-style: solid;, border-radius: 12px; & finally align-items: center;