changes to style.css for the quote container
I added a comment for what to do later in regards to the subtext under the images.

I added a class selector .quote-container and an ID selector #author-quote.

For .quote-container I did display: flex; with direction column; centered the content, background-color: #E5E7EB with a font-size: 36px;, font-weight: bold, and a padding of 50px;.

For the ID selector in order to make on the right side under the quote I made it display flex but justify-content: flex-end;.