signin-header & subtext changes
In index.html I changed the Id to signin instead of signing.

I also added an ID attribute to the p element under the ID signin-header. That is all the changes made to index.html

For style.css I added the property margin-bottom with value 0 to signin-header also I changed the id selector to the new updated name as well.

I also added a new id selector signin-subtext with margin-top:0;

That is all the changes I made.