Finaly rough draft to project
Finished doing the footer of the webpage by center content and turning it into a flexbox.
I also gave it a background color as I did to the heade. I will go back and try to do little changes to make the website stand out and more functional. Like adding italicize to the quote and font-styles as well.